:bug: Provide log path to expected outputs logger
